[Design] : set locale in daysOfWeekStyle
Git : #32
